requirements.txt
macfsevents
fabric
